Add home page banner & move element styles
This adds an image and styles for the home page banner, as well as
extracts the reusable elements styles into their own stylesheet so they
are easier to manage.

The banner is not yet full width, but that may come in the future based
on feedback and thoughts. Currently it is constrained to the width of
the wrapper.