Add full-width page layout for pages w/ banners
This makes it so that there is a new layout, fw_page, that can be used
for pages that need a full-width custom layout. Specifically, it is for
use on the Home & About page where the banners need to be full-width
instead of constrained within the wrapper.

Pages that use the fw_page layout are responsible for their own mark-up
to have certain content within different wrappers.

This updates the other existing layouts to have the wrapper set so their
layouts do not change.