Increase nav title & link size
This increases the height of the site header nav to make room for a
larger logo and site title, as well as some large nav link items. There
is also a style adjustment for the nav wrapper to use flexbox to
vertically center it within the nav.

This change was made based on feedback from Peter S. prior to making the
site live.

PT Story: [#136159475](https://www.pivotaltracker.com/story/show/136159475)