Update Contact page content & layout
This updates the contact email and site config for the new domain
that'll be used, savilabs.org, and it removes the placeholder address
from the contact page. Now the Contact page content is displayed
centered to make the most of the available space.

This change was made in preparation for launching the site.