Tweak text spacing for featured content
This makes it so that the home page LaunchPad features list displays as
a bulleted list, as well as tweaks text spacing so that it is more
consistent throughout the site. I made these changes based on feedback
from Peter and some styles I noticed that weren't sticking.

PT Story: [#129631153](https://www.pivotaltracker.com/story/show/129631153)